Vulkan Pixel History: Add a check for multiple subpasses
This primarily so we don't crash on renderpasses with multiple
subpasses, they are still not fully supported.
With this change pixel history will give the events that passed
occlusion query and indicate which test events failed (if any).
It will not return any colour, depth, stencil, shader out information,
or give the number of fragments in the event.